Create the initial CiscoSparkClient class interface
Create the intial CiscoSparkClient class interface and initialize the contained CiscoSparkAPI connection object and expose it as `CiscoSparkClient.api`.

Ensure the interface passes all of the previously defined TDD test suite, and define an additional tests to ensure the CiscoSparkAPI object is properly initialized and that the access token is accessible via the CiscoSparkClient interface.